+++ sshubl/ssh_utils.py
@@ -87,9 +87,9 @@ def ssh_connect(
     # run OpenSSH using pexpect to setup connection and non-interactively deal with prompts
     ssh = pxssh.pxssh(
         options={
-            **(_settings().get("ssh_options") or {}),
+            **_settings().get("ssh_options", {}),
             # enforce keep-alive for future sshfs usages (see upstream recommendations)
-            "ServerAliveInterval": str((_settings().get("ssh_server_alive_interval") or 15)),
+            "ServerAliveInterval": str(_settings().get("ssh_server_alive_interval", 15)),
             "ControlMaster": "auto",
             "ControlPath": str(sockets_path / str(identifier)),
             # keep connection opened for 1 minute (without new connection to control socket)
@@ -110,12 +110,12 @@ def ssh_connect(
             host,
             login,
             password or "",
-            login_timeout=_settings().get("ssh_login_timeout") or 10,
+            login_timeout=_settings().get("ssh_login_timeout", 10),
             port=port,
             auto_prompt_reset=False,
             cmd=ssh_program,
             # allow user to disable host authentication for loopback addresses
-            check_local_ip=_settings().get("ssh_host_authentication_for_localhost") or True,
+            check_local_ip=_settings().get("ssh_host_authentication_for_localhost", True),
         )
     except pxssh.ExceptionPxssh as exception:
         # if authentication failed without password, raise a specific exception