+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.0] - 2025-08-06
+
+### Added
+
+- Configurable OpenSSH control sockets path directory
+
 ## [0.5.0] - 2024-11-30
 
 ### Added
@@ -77,7 +83,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/HorlogeSkynet/SSHubl/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/HorlogeSkynet/SSHubl/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/HorlogeSkynet/SSHubl/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/HorlogeSkynet/SSHubl/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/HorlogeSkynet/SSHubl/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/HorlogeSkynet/SSHubl/compare/v0.2.1...v0.3.0